refactor(eslint): split dependency contract rules
