+++ eslint-local-rules/enforce-thunk-contracts/rule.ts
@@ -1,29 +1,18 @@
 import type { Rule } from 'eslint';
 import ts from 'typescript';
 
-type TypeAwareParserServices = {
-    esTreeNodeToTSNodeMap?: ReadonlyMap<Rule.Node, ts.Node>;
-    program?: ts.Program;
-    tsNodeToESTreeNodeMap?: ReadonlyMap<ts.Node, Rule.Node>;
-};
+import {
+    createNamedContractRuleListener,
+    getFunctionExpression,
+    getTypeReferenceName,
+    getVariableFunction,
+    unwrapExpression,
+} from '../named-contracts/utils';
 
 type RequiredContract = {
     expectedName: string;
     node: ts.TypeNode;
 };
-
-type MessageId =
-    | 'contractMustBeAdjacent'
-    | 'contractMustBeSeparated'
-    | 'contractOrder'
-    | 'contractMustBeNamed'
-    | 'dependencyFactoryContractOrder'
-    | 'dependencyFactoryParameter'
-    | 'dependencyFactoryReturnType'
-    | 'emptyThunkConfig'
-    | 'missingThunkConfig'
-    | 'thunkConfigMustBeInline'
-    | 'voidContractProperty';
 
 const thunkCreators = new Set(['createSingleInstanceThunk', 'createThunk']);
 
@@ -41,27 +30,6 @@ const getThunkContractBaseName = (name: string) => {
     }
 
     return capitalizedName.endsWith('Thunk') ? capitalizedName : `${capitalizedName}Thunk`;
-};
-
-const unwrapExpression = (expression: ts.Expression): ts.Expression => {
-    if (
-        ts.isParenthesizedExpression(expression) ||
-        ts.isAsExpression(expression) ||
-        ts.isTypeAssertionExpression(expression) ||
-        ts.isNonNullExpression(expression)
-    ) {
-        return unwrapExpression(expression.expression);
-    }
-
-    return expression;
-};
-
-const getFunctionExpression = (expression: ts.Expression) => {
-    const unwrappedExpression = unwrapExpression(expression);
-
-    return ts.isArrowFunction(unwrappedExpression) || ts.isFunctionExpression(unwrappedExpression)
-        ? unwrappedExpression
-        : undefined;
 };
 
 const getReturnedFunction = (functionNode: ts.FunctionLikeDeclaration) => {
@@ -95,11 +63,6 @@ const getInnermostReturnedFunction = (functionNode: ts.FunctionLikeDeclaration) 
     return currentFunction;
 };
 
-const getTypeReferenceName = (node: ts.TypeNode | undefined) =>
-    node !== undefined && ts.isTypeReferenceNode(node) && ts.isIdentifier(node.typeName)
-        ? node.typeName.text
-        : undefined;
-
 const getPropertyName = (node: ts.TypeElement) => {
     if (!ts.isPropertySignature(node) || node.name === undefined) {
         return undefined;
@@ -127,11 +90,6 @@ const isEmptyContractType = (node: ts.TypeNode | undefined) => {
         node.typeArguments?.[0]?.kind === ts.SyntaxKind.NeverKeyword
     );
 };
-
-const getVariableFunction = (declaration: ts.VariableDeclaration) =>
-    declaration.initializer === undefined
-        ? undefined
-        : getFunctionExpression(declaration.initializer);
 
 const getThunkImplementationFromFunction = (outerFunction: ts.FunctionLikeDeclaration) => {
     const implementation = getInnermostReturnedFunction(outerFunction);
@@ -162,15 +120,12 @@ const getGetStateReturnType = (parameter: ts.ParameterDeclaration | undefined) =
     return ts.isFunctionTypeNode(parameter.type) ? parameter.type.type : parameter.type;
 };
 
-/**
- * Enforces local, predictably named state and dependency contracts for thunks and DI factories.
- */
-export const enforceNamedContractsRule: Rule.RuleModule = {
+/** Enforces local, predictably named state and dependency contracts for thunks. */
+export const enforceThunkContractsRule: Rule.RuleModule = {
     meta: {
         type: 'problem',
         docs: {
-            description:
-                'Enforces explicit, named State and Deps contracts for thunks and dependency-injected service factories.',
+            description: 'Enforces explicit, named State and Deps contracts for thunks.',
             category: 'Best Practices',
             recommended: false,
         },
@@ -183,12 +138,6 @@ export const enforceNamedContractsRule: Rule.RuleModule = {
             contractOrder: "Declare '{{stateName}}' before '{{depsName}}' for '{{consumerName}}'.",
             contractMustBeNamed:
                 "Use the named '{{contractName}}' contract instead of an inline or differently named type.",
-            dependencyFactoryParameter:
-                "Pass '{{contractName}}' as a single 'deps' parameter to dependency-injected factory '{{factoryName}}'.",
-            dependencyFactoryContractOrder:
-                "Declare '{{depsName}}' before '{{serviceName}}' for '{{factoryName}}'.",
-            dependencyFactoryReturnType:
-                "Give dependency-injected factory '{{factoryName}}' an explicit named service return type.",
             emptyThunkConfig: "Use explicit 'void' for dependency-free thunk '{{thunkName}}'.",
             missingThunkConfig:
                 "Give thunk '{{thunkName}}' an explicit third generic: 'void' or a named State/Deps config.",
@@ -199,154 +148,17 @@ export const enforceNamedContractsRule: Rule.RuleModule = {
         },
         schema: [],
     },
-    create(context) {
-        const parserServices = context.sourceCode.parserServices as TypeAwareParserServices;
-
-        if (
-            parserServices.program === undefined ||
-            parserServices.esTreeNodeToTSNodeMap === undefined ||
-            parserServices.tsNodeToESTreeNodeMap === undefined
-        ) {
-            return {};
-        }
-
-        return {
-            'Program:exit': programNode => {
-                const sourceFile = parserServices.esTreeNodeToTSNodeMap?.get(
-                    programNode as Rule.Node,
-                );
-
-                if (sourceFile === undefined || !ts.isSourceFile(sourceFile)) {
-                    return;
-                }
-
-                const { statements } = sourceFile;
-                const localTypeAliases = new Map(
-                    statements
-                        .filter(ts.isTypeAliasDeclaration)
-                        .map(declaration => [declaration.name.text, declaration] as const),
-                );
-
-                const report = (
-                    node: ts.Node,
-                    messageId: MessageId,
-                    data: Record<string, string>,
-                    fix?: Rule.ReportFixer,
-                ) => {
-                    const reportNode = parserServices.tsNodeToESTreeNodeMap?.get(node);
-
-                    if (reportNode !== undefined) {
-                        context.report({ node: reportNode, messageId, data, fix });
-                    }
-                };
-
-                const createAdjacentDeclarationSwapFix = (
-                    firstDeclaration: ts.Statement,
-                    secondDeclaration: ts.Statement,
-                ): Rule.ReportFixer | undefined => {
-                    const firstStatementIndex = statements.indexOf(firstDeclaration);
-                    const secondStatementIndex = statements.indexOf(secondDeclaration);
-
-                    if (secondStatementIndex !== firstStatementIndex + 1) {
-                        return undefined;
-                    }
-
-                    const previousStatement = statements[firstStatementIndex - 1];
-                    const nextStatement = statements[secondStatementIndex + 1];
-                    const firstStart = firstDeclaration.getStart(sourceFile);
-                    const secondStart = secondDeclaration.getStart(sourceFile);
-                    const leadingTrivia = sourceFile.text.slice(
-                        previousStatement?.end ?? 0,
-                        firstStart,
-                    );
-                    const betweenDeclarations = sourceFile.text.slice(
-                        firstDeclaration.end,
-                        secondStart,
-                    );
-                    const trailingTrivia = sourceFile.text.slice(
-                        secondDeclaration.end,
-                        nextStatement?.getStart(sourceFile) ?? sourceFile.end,
-                    );
-                    const containsComment = (text: string) => /\/\/|\/\*/u.test(text);
-
-                    // A comment in the surrounding whitespace may describe one declaration.
-                    // Moving code without knowing which declaration owns it could change its meaning.
-                    if (
-                        containsComment(leadingTrivia) ||
-                        containsComment(betweenDeclarations) ||
-                        containsComment(trailingTrivia)
-                    ) {
-                        return undefined;
-                    }
-
-                    const firstText = sourceFile.text.slice(firstStart, firstDeclaration.end);
-                    const secondText = sourceFile.text.slice(secondStart, secondDeclaration.end);
-
-                    return fixer =>
-                        fixer.replaceTextRange(
-                            [firstStart, secondDeclaration.end],
-                            `${secondText}${betweenDeclarations}${firstText}`,
-                        );
-                };
-
-                const validateContractBlockSpacing = (
-                    firstStatementIndex: number,
-                    consumerStatementIndex: number,
-                    consumerName: string,
-                ) => {
-                    const contractBlock = statements.slice(
-                        firstStatementIndex,
-                        consumerStatementIndex + 1,
-                    );
-
-                    for (let index = 1; index < contractBlock.length; index += 1) {
-                        const previousStatement = contractBlock[index - 1];
-                        const nextStatement = contractBlock[index];
-
-                        if (previousStatement === undefined || nextStatement === undefined) {
-                            continue;
-                        }
-
-                        const textBetweenStatements = sourceFile.text.slice(
-                            previousStatement.end,
-                            nextStatement.getStart(sourceFile),
-                        );
-
-                        if (/\r?\n[\t ]*\r?\n/u.test(textBetweenStatements)) {
-                            continue;
-                        }
-
-                        const previousName =
-                            (ts.isTypeAliasDeclaration(previousStatement) ||
-                                ts.isInterfaceDeclaration(previousStatement)) &&
-                            previousStatement.name.text;
-                        const nextName =
-                            (ts.isTypeAliasDeclaration(nextStatement) ||
-                                ts.isInterfaceDeclaration(nextStatement)) &&
-                            nextStatement.name.text;
-                        const reportNode = parserServices.tsNodeToESTreeNodeMap?.get(nextStatement);
-
-                        if (reportNode !== undefined) {
-                            context.report({
-                                node: reportNode,
-                                messageId: 'contractMustBeSeparated',
-                                data: {
-                                    previousName: previousName || 'the previous declaration',
-                                    nextName: nextName || consumerName,
-                                },
-                                fix: fixer => {
-                                    const { line } = sourceFile.getLineAndCharacterOfPosition(
-                                        nextStatement.getStart(sourceFile),
-                                    );
-                                    const start = sourceFile.getPositionOfLineAndCharacter(line, 0);
-
-                                    return fixer.insertTextBeforeRange([start, start], '\n');
-                                },
-                            });
-                        }
-                    }
-                };
-
+    create: context =>
+        createNamedContractRuleListener(
+            context,
+            ({
+                createAdjacentDeclarationSwapFix,
+                localTypeAliases,
+                report,
+                sourceFile,
+                statements,
+                validateContractBlockSpacing,
+            }) => {
                 const validateNamedContract = (
                     node: ts.TypeNode | undefined,
                     expectedName: string,
@@ -556,121 +368,15 @@ export const enforceNamedContractsRule: Rule.RuleModule = {
                     validateAdjacentContracts(statementIndex, thunkName, requiredContracts);
                 };
 
-                const validateDependencyFactory = (
-                    factoryName: string,
-                    factoryFunction: ts.FunctionLikeDeclaration,
-                    statementIndex: number,
-                ) => {
-                    // Composition roots and dependency-object builders wire services together;
-                    // they are not factories for the service named after `create`.
-                    if (factoryName.endsWith('CompositionRoot') || factoryName.endsWith('Deps')) {
-                        return;
-                    }
-
-                    const depsParameter = factoryFunction.parameters[0];
-
-                    if (depsParameter === undefined) {
-                        return;
-                    }
-
-                    const serviceName = factoryName.slice('create'.length);
-                    const expectedDepsName = `${serviceName}Deps`;
-                    const actualDepsName = getTypeReferenceName(depsParameter.type);
-                    const isNamedDepsParameter =
-                        ts.isIdentifier(depsParameter.name) && depsParameter.name.text === 'deps';
-
-                    if (!isNamedDepsParameter && !actualDepsName?.endsWith('Deps')) {
-                        return;
-                    }
-
-                    if (!isNamedDepsParameter) {
-                        report(depsParameter.name, 'dependencyFactoryParameter', {
-                            factoryName,
-                            contractName: expectedDepsName,
-                        });
-                    }
-
-                    if (actualDepsName !== expectedDepsName) {
-                        report(depsParameter.type ?? depsParameter, 'contractMustBeNamed', {
-                            consumerName: factoryName,
-                            contractName: expectedDepsName,
-                        });
-                    }
-
-                    const returnTypeName = getTypeReferenceName(factoryFunction.type);
-
-                    if (returnTypeName === undefined) {
-                        report(factoryFunction, 'dependencyFactoryReturnType', { factoryName });
-                    }
-
-                    const depsDeclaration = localTypeAliases.get(expectedDepsName);
-
-                    if (depsDeclaration !== undefined) {
-                        const depsStatementIndex = statements.indexOf(depsDeclaration);
-                        const serviceDeclaration =
-                            returnTypeName === undefined
-                                ? undefined
-                                : localTypeAliases.get(returnTypeName);
-                        const serviceStatementIndex =
-                            serviceDeclaration === undefined
-                                ? undefined
-                                : statements.indexOf(serviceDeclaration);
-                        const firstContractIndex = Math.min(
-                            depsStatementIndex,
-                            serviceStatementIndex ?? depsStatementIndex,
-                        );
-                        const isInAdjacentTypeBlock = statements
-                            .slice(firstContractIndex, statementIndex)
-                            .every(
-                                statement =>
-                                    ts.isTypeAliasDeclaration(statement) ||
-                                    ts.isInterfaceDeclaration(statement),
-                            );
-
-                        if (isInAdjacentTypeBlock) {
-                            if (
-                                serviceDeclaration !== undefined &&
-                                serviceStatementIndex !== undefined &&
-                                depsStatementIndex > serviceStatementIndex
-                            ) {
-                                report(
-                                    depsDeclaration,
-                                    'dependencyFactoryContractOrder',
-                                    {
-                                        factoryName,
-                                        depsName: expectedDepsName,
-                                        serviceName: serviceDeclaration.name.text,
-                                    },
-                                    createAdjacentDeclarationSwapFix(
-                                        serviceDeclaration,
-                                        depsDeclaration,
-                                    ),
-                                );
-                            }
-
-                            validateContractBlockSpacing(
-                                firstContractIndex,
-                                statementIndex,
-                                factoryName,
-                            );
-                        }
-                    }
-                };
-
                 statements.forEach((statement, statementIndex) => {
                     if (ts.isFunctionDeclaration(statement)) {
                         if (statement.name !== undefined) {
-                            const declarationName = statement.name.text;
                             const vanillaThunk = getThunkImplementationFromFunction(statement);
 
                             if (vanillaThunk !== undefined) {
-                                validateVanillaThunk(declarationName, vanillaThunk, statementIndex);
-                            }
-
-                            if (declarationName.startsWith('create')) {
-                                validateDependencyFactory(
-                                    declarationName,
-                                    statement,
+                                validateVanillaThunk(
+                                    statement.name.text,
+                                    vanillaThunk,
                                     statementIndex,
                                 );
                             }
@@ -708,21 +414,8 @@ export const enforceNamedContractsRule: Rule.RuleModule = {
                         if (vanillaThunk !== undefined) {
                             validateVanillaThunk(declarationName, vanillaThunk, statementIndex);
                         }
-
-                        if (declarationName.startsWith('create')) {
-                            const factoryFunction = getVariableFunction(declaration);
-
-                            if (factoryFunction !== undefined) {
-                                validateDependencyFactory(
-                                    declarationName,
-                                    factoryFunction,
-                                    statementIndex,
-                                );
-                            }
-                        }
                     }
                 });
             },
-        };
-    },
+        ),
 };
