+++ eslint-local-rules/index.js
@@ -6,7 +6,8 @@ require('ts-node').register({
 });
 
 const { analyticsEventNameRule } = require('./analytics-event-name/rule');
-const { enforceNamedContractsRule } = require('./enforce-named-contracts/rule');
+const { enforceDiFactoryContractsRule } = require('./enforce-di-factory-contracts/rule');
+const { enforceThunkContractsRule } = require('./enforce-thunk-contracts/rule');
 const { noOverrideDsComponentRule } = require('./no-override-ds-component/rule');
 const { noPackageDeepImportsRule } = require('./no-package-deep-imports/rule');
 const { noSuiteImportsInSuiteCommonRule } = require('./no-suite-imports-in-suite-common/rule');
@@ -14,7 +15,8 @@ const { noUnusedIntersectionMembersRule } = require('./no-unused-intersection-me
 
 module.exports = {
     'analytics-event-name': analyticsEventNameRule,
-    'enforce-named-contracts': enforceNamedContractsRule,
+    'enforce-di-factory-contracts': enforceDiFactoryContractsRule,
+    'enforce-thunk-contracts': enforceThunkContractsRule,
     'no-override-ds-component': noOverrideDsComponentRule,
     'no-package-deep-imports': noPackageDeepImportsRule,
     'no-suite-imports-in-suite-common': noSuiteImportsInSuiteCommonRule,
